Fix GET parameters handling
The current GET parameters handling is incorrect and not needed. The
issue is with parameters that include only a key and no value. This is
common with WSDL services that often include a value-less parameter
called "wsdl". The current code assumes there's a value, so it adds the
key to the URL with an equality sign. This can cause the web service to
fail and we won't get the original response.
The solution is really simple, just remove this piece of code.
To see this in action, try to proxy the following URL with and without
the fix: http://wsf.cdyne.com/WeatherWS/Weather.asmx?WSDL