New feature: URL whitelist
Currently when a user uses the service, he can enter any URL he wants.
For certain projects this might be an issue because of possible abuse.
I added a whitelist mechanism to allow the developer to configure what
URLs he wants to proxy. For that we have a new file called config.php
which contains an array that you can fill with regex patterns matching
the URLs you want to accept. If you leave it empty, it will allow any
URL, just like it did before.
See config.php for more info and examples in the comments.