New feature: add CORS support
When a developer is working on a website/webapp that is accessing
miniProxy using AJAX, the developer might have issues with CORS
(Cross-origin resource sharing) that will prevent him/her from getting
the proxified data.
I added a way to easily enable CORS support. All the developer needs to
do is edit config.php and set $forceCORS to true (the default is false,
which will act exactly as before).
Using this will make it a lot easier to use the service through AJAX.